class loader problem mit jaiio

+++ org.jmedikit.plugin/src/org/jmedikit/lib/image/AbstractImage.java
@@ -254,7 +254,7 @@ public abstract class AbstractImage {
 		short minValue = Short.MAX_VALUE;
 		
 		for(int i = 0; i < pixels.length; i++){
-			short value = (short)pixels[i];
+			short value = pixels[i];
 			if(value < minValue){
 				minValue = value;
 			}
